chore: bump packages (#236)

+++ chains/substrate/config.go
@@ -31,7 +31,7 @@ type SubstrateConfig struct {
 	StartBlock         *big.Int
 	BlockInterval      *big.Int
 	BlockRetryInterval time.Duration
-	SubstrateNetwork   uint8
+	SubstrateNetwork   uint16
 	Tip                uint64
 }
 
@@ -90,7 +90,7 @@ func NewSubstrateConfig(chainConfig map[string]interface{}) (*SubstrateConfig, e
 		BlockRetryInterval: time.Duration(c.BlockRetryInterval) * time.Second,
 		StartBlock:         big.NewInt(c.StartBlock),
 		BlockInterval:      big.NewInt(c.BlockInterval),
-		SubstrateNetwork:   uint8(c.SubstrateNetwork),
+		SubstrateNetwork:   uint16(c.SubstrateNetwork),
 		Tip:                uint64(c.Tip),
 	}
