+++ cli/utils/derivateSS58AccountFromPK.go
@@ -5,6 +5,7 @@ package utils
 
 import (
 	"fmt"
+
 	"github.com/centrifuge/go-substrate-rpc-client/v4/signature"
 	"github.com/spf13/cobra"
 )
@@ -20,13 +21,13 @@ var (
 
 var (
 	privateKey string
-	networkID  uint8
+	networkID  uint16
 )
 
 func init() {
 	derivateSS58AccountFromPKCMD.PersistentFlags().StringVar(&privateKey, "privateKey", "", "hex encoded private key")
 	_ = derivateSS58AccountFromPKCMD.MarkFlagRequired("privateKey")
-	derivateSS58AccountFromPKCMD.PersistentFlags().Uint8Var(&networkID, "networkID", 0, "network id for a checksum. Registry https://github.com/paritytech/ss58-registry/blob/main/ss58-registry.json")
+	derivateSS58AccountFromPKCMD.PersistentFlags().Uint16Var(&networkID, "networkID", 0, "network id for a checksum. Registry https://github.com/paritytech/ss58-registry/blob/main/ss58-registry.json")
 	_ = derivateSS58AccountFromPKCMD.MarkFlagRequired("networkID")
 }
